requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
websockets==15.0.1
dotenv==0.9.9
google-genai==1.52.0